change performance form types namespace
