+++ src/PrestaShopBundle/Form/Admin/Configure/AdvancedParameters/Performance/MediaServersType.php
@@ -23,16 +23,16 @@
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  * International Registered Trademark & Property of PrestaShop SA
  */
-namespace PrestaShopBundle\Form\Admin\AdvancedParameters\Performance;
+namespace PrestaShopBundle\Form\Admin\Configure\AdvancedParameters\Performance;
 
 use PrestaShopBundle\Form\Admin\Type\CommonAbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
 /**
- * This form class generates the "Memcache server" form in Performance page
+ * This form class generates the "Media servers" form in Performance page
  */
-class MemcacheServerType extends CommonAbstractType
+class MediaServersType extends CommonAbstractType
 {
     /**
      * {@inheritdoc}
@@ -40,13 +40,13 @@ class MemcacheServerType extends CommonAbstractType
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder
-            ->add('memcache_ip', 'Symfony\Component\Form\Extension\Core\Type\TextType', array(
+            ->add('media_server_one', 'Symfony\Component\Form\Extension\Core\Type\TextType', array(
                 'required' => false,
             ))
-            ->add('memcache_port', 'Symfony\Component\Form\Extension\Core\Type\TextType', array(
+            ->add('media_server_two', 'Symfony\Component\Form\Extension\Core\Type\TextType', array(
                 'required' => false,
             ))
-            ->add('memcache_weight', 'Symfony\Component\Form\Extension\Core\Type\TextType', array(
+            ->add('media_server_three', 'Symfony\Component\Form\Extension\Core\Type\TextType', array(
                 'required' => false,
             ))
         ;
@@ -67,6 +67,6 @@ class MemcacheServerType extends CommonAbstractType
      */
     public function getBlockPrefix()
     {
-        return 'performance_memcache_server_block';
+        return 'performance_media_servers_block';
     }
 }
