+++ src/PrestaShopBundle/Form/Admin/Configure/AdvancedParameters/Performance/MemcacheServerType.php
@@ -23,18 +23,16 @@
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  * International Registered Trademark & Property of PrestaShop SA
  */
-namespace PrestaShopBundle\Form\Admin\AdvancedParameters\Performance;
+namespace PrestaShopBundle\Form\Admin\Configure\AdvancedParameters\Performance;
 
 use PrestaShopBundle\Form\Admin\Type\CommonAbstractType;
-use PrestaShopBundle\Form\Admin\Type\SwitchType;
-use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
 /**
- * This form class generates the "Smarty" form in Performance page
+ * This form class generates the "Memcache server" form in Performance page
  */
-class SmartyType extends CommonAbstractType
+class MemcacheServerType extends CommonAbstractType
 {
     /**
      * {@inheritdoc}
@@ -42,33 +40,14 @@ class SmartyType extends CommonAbstractType
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder
-            ->add('template_compilation', ChoiceType::class, array(
-                'choices'  => array(
-                    'Never recompile template files' => 0,
-                    'Recompile templates if the files have been updated' => 1,
-                    'Force compilation' => 2,
-                ),
-                'required' => true,
+            ->add('memcache_ip', 'Symfony\Component\Form\Extension\Core\Type\TextType', array(
+                'required' => false,
             ))
-            ->add('cache', SwitchType::class, array(
-                'required' => true,
+            ->add('memcache_port', 'Symfony\Component\Form\Extension\Core\Type\TextType', array(
+                'required' => false,
             ))
-            ->add('multi_front_optimization', SwitchType::class, array(
-                'required' => true,
-            ))
-            ->add('caching_type', ChoiceType::class, array(
-                'choices'  => array(
-                    'File System' => 'filesystem',
-                    'MySQL' => 'mysql',
-                ),
-                'required' => true,
-            ))
-            ->add('clear_cache', ChoiceType::class, array(
-                'choices'  => array(
-                     'Never clear cache files' => 'never',
-                    'Clear cache everytime something has been modified' => 'everytime',
-                ),
-                'required' => true,
+            ->add('memcache_weight', 'Symfony\Component\Form\Extension\Core\Type\TextType', array(
+                'required' => false,
             ))
         ;
     }
@@ -88,6 +67,6 @@ class SmartyType extends CommonAbstractType
      */
     public function getBlockPrefix()
     {
-        return 'performance_smarty_block';
+        return 'performance_memcache_server_block';
     }
 }
