+++ src/PrestaShopBundle/Form/Admin/Configure/AdvancedParameters/Performance/PerformanceFormDataProvider.php
@@ -23,7 +23,7 @@
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  * International Registered Trademark & Property of PrestaShop SA
  */
-namespace PrestaShopBundle\Form\Admin\AdvancedParameters\Performance;
+namespace PrestaShopBundle\Form\Admin\Configure\AdvancedParameters\Performance;
 
 use PrestaShop\PrestaShop\Adapter\OptionalFeatures\OptionalFeaturesConfiguration;
 use PrestaShop\PrestaShop\Adapter\Cache\CombineCompressCacheConfiguration;
