+++ src/PrestaShopBundle/Resources/config/services/form/form_type.yml
@@ -238,38 +238,38 @@ services:
             - "@=service('prestashop.adapter.legacy.context').getLanguages()"
 
     form.type.performance.smarty:
-        class: 'PrestaShopBundle\Form\Admin\AdvancedParameters\Performance\SmartyType'
+        class: 'PrestaShopBundle\Form\Admin\Configure\AdvancedParameters\Performance\SmartyType'
         tags:
             - { name: form.type }
 
     form.type.performance.debug_mode:
-        class: 'PrestaShopBundle\Form\Admin\AdvancedParameters\Performance\DebugModeType'
+        class: 'PrestaShopBundle\Form\Admin\Configure\AdvancedParameters\Performance\DebugModeType'
         tags:
             - { name: form.type }
 
     form.type.performance.optional_features:
-        class: 'PrestaShopBundle\Form\Admin\AdvancedParameters\Performance\OptionalFeaturesType'
+        class: 'PrestaShopBundle\Form\Admin\Configure\AdvancedParameters\Performance\OptionalFeaturesType'
         arguments: ['@=service("prestashop.adapter.combination_feature").isUsed()']
         tags:
             - { name: form.type }
 
     form.type.performance.ccc:
-        class: 'PrestaShopBundle\Form\Admin\AdvancedParameters\Performance\CombineCompressCacheType'
+        class: 'PrestaShopBundle\Form\Admin\Configure\AdvancedParameters\Performance\CombineCompressCacheType'
         tags:
             - { name: form.type }
 
     form.type.performance.media_servers:
-        class: 'PrestaShopBundle\Form\Admin\AdvancedParameters\Performance\MediaServersType'
+        class: 'PrestaShopBundle\Form\Admin\Configure\AdvancedParameters\Performance\MediaServersType'
         tags:
             - { name: form.type }
 
     form.type.performance.memcache_servers:
-        class: 'PrestaShopBundle\Form\Admin\AdvancedParameters\Performance\MemcacheServerType'
+        class: 'PrestaShopBundle\Form\Admin\Configure\AdvancedParameters\Performance\MemcacheServerType'
         tags:
             - { name: form.type }
 
     form.type.performance.caching:
-        class: 'PrestaShopBundle\Form\Admin\AdvancedParameters\Performance\CachingType'
+        class: 'PrestaShopBundle\Form\Admin\Configure\AdvancedParameters\Performance\CachingType'
         parent: 'form.type.translatable.aware'
         public: true
         tags:
